+++ integration/crud-typeorm/companies/responses/index.ts
@@ -1,4 +1,4 @@
-import { SerializeOptions } from '@nestjsx/crud';
+import { SerializeOptions } from '@recalibratedsystems/nestjs-crud';
 import { GetCompanyResponseDto } from './get-company-response.dto';
 
 export const serialize: SerializeOptions = {