+++ packages/crud-request/src/interfaces/parsed-request.interface.ts
@@ -1,4 +1,4 @@
-import { ObjectLiteral } from '@recalibratedsystems/netsjs-crud-util';
+import { ObjectLiteral } from '@recalibratedsystems/nestjs-crud-util';
 import { QueryFields, QueryFilter, QueryJoin, QuerySort, SCondition } from '../types';
 
 export interface ParsedRequestParams {