+++ packages/crud-request/src/request-query.parser.ts
@@ -10,7 +10,7 @@ import {
   objKeys,
   isNil,
   ObjectLiteral,
-} from '@recalibratedsystems/netsjs-crud-util';
+} from '@recalibratedsystems/nestjs-crud-util';
 
 import { RequestQueryException } from './exceptions';
 import { ParamsOptions, ParsedRequestParams, RequestQueryBuilderOptions } from './interfaces';