+++ packages/crud/src/crud/reflection.helper.ts
@@ -8,7 +8,7 @@ import {
   ROUTE_ARGS_METADATA,
 } from '@nestjs/common/constants';
 import { ArgumentsHost } from '@nestjs/common';
-import { isFunction } from '@recalibratedsystems/netsjs-crud-util';
+import { isFunction } from '@recalibratedsystems/nestjs-crud-util';
 
 import { BaseRoute, MergedCrudOptions, AuthOptions } from '../interfaces';
 import { BaseRouteName } from '../types';